Simplify linha quente bifurcation logging
- Remove separate no-bifurcation reason from debug output
- Reuse linha fria reason when hot path converges or refuses
- Update tests and shared fixtures for the new contract
